+++ R/read_structure.R
@@ -168,7 +168,7 @@ part_style_list <- function(x, fun = NULL, more_args = list()){
 par_style_list <- function(x){
 
   fp_columns <- intersect(names(formals(officer::fp_par)), colnames(x))
-  dat <- x[c(fp_columns, "text.direction", "vertical.align",
+  dat <- as.data.frame(x)[c(fp_columns, "text.direction", "vertical.align",
              grep("^border\\.", colnames(x), value = TRUE))]
   setDT(dat)
   uid <- unique(dat)
@@ -208,8 +208,8 @@ cell_style_list <- function(x){
 
   fp_columns <- intersect(names(formals(officer::fp_cell)), colnames(x))
 
-  dat <- x[c(fp_columns, "text.align", "width", "height", "hrule", grep("^border\\.", colnames(x), value = TRUE))]
-  setDT(dat)
+  dat <- as.data.frame(x)
+  dat <- dat[c(fp_columns, "text.align", "width", "height", "hrule", grep("^border\\.", colnames(dat), value = TRUE))]
 
   uid <- unique(dat)
   classname <- UUIDgenerate(n = nrow(uid), use.time = TRUE)